# swagger.yaml
openapi: 3.0.0
info:
  title: SubastaPlus API
  description: |
    API REST para la aplicación móvil SubastaPlus - Sistema de Subastas.
    Trabajo Práctico Obligatorio - Desarrollo de Aplicaciones I - 1C2026.



    ## Roles
    - **invitado**: usuario no autenticado, acceso limitado a catálogos públicos (sin precios base)
    - **postor**: usuario registrado y aprobado, puede pujar según su categoría
  version: 1.0.0
  contact:
    name: Equipo SubastaPlus
servers:
  - description: SwaggerHub API Auto Mocking
    url: 'https://virtserver.swaggerhub.com/uade-05d/TPO-DesarrolloDeAps/1.0.0'
  - url: 'https://api.subastaplus.com/v1'
    description: Producción
  - url: 'http://localhost:3000/v1'
    description: Desarrollo local
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CategoriaUsuario:
      type: string
      enum:
        - comun
        - especial
        - plata
        - oro
        - platino
...
    SubastaResumen:
      type: object
      properties:
        id:
          type: string
          format: uuid
        titulo:
          type: string
          example: 'Arte Moderno #47'
        categoria:
          $ref: '#/components/schemas/CategoriaUsuario'
        fecha:
          type: string
          format: date-time
        moneda:
          type: string
          enum:
            - ARS
            - USD
        estado:
          type: string
          enum:
            - programada
            - en_vivo
            - finalizada
        cantidadPiezas:
          type: integer
          example: 12
        rematador:
          type: string
        esColeccion:
          type: boolean
        nombreColeccion:
          type: string
          nullable: true
...
    SolicitudVenta:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tipo:
          type: string
          enum:
            - arte
            - antiguedad
            - joya
            - vehiculo
            - mueble
            - otro
        descripcion:
          type: string
        imagenes:
          type: array
          items:
            type: string
            format: uri
          minItems: 6
        historia:
          type: string
          nullable: true
        declaracionPropiedad:
          type: boolean
        estado:
          type: string
          enum:
            - borrador
            - enviada
            - en_revision
            - aceptada
            - rechazada
            - en_subasta
            - vendida
            - no_vendida
        motivoRechazo:
          type: string
          nullable: true
        valorBase:
          type: number
          nullable: true
        comisiones:
          type: number
          nullable: true
        subastaAsignada:
          $ref: '#/components/schemas/SubastaResumen'
        direccionEnvio:
          type: string
          nullable: true
          description: Dirección a la que el usuario debe enviar el bien
        ubicacionDeposito:
          type: string
          nullable: true
        polizaSeguro:
          $ref: '#/components/schemas/PolizaSeguro'
        cuentaCobro:
          type: string
          nullable: true
          description: Cuenta donde se depositará el dinero de la venta
        fechaCreacion:
          type: string
          format: date-time
...
    PolizaSeguro:
      type: object
      nullable: true
      properties:
        id:
          type: string
          format: uuid
        aseguradora:
          type: string
        numeroPoliza:
          type: string
        valorAsegurado:
          type: number
        contactoAseguradora:
          type: string
          description: Teléfono o email de la aseguradora
...
paths:
...
  '/solicitudes-venta/{id}':
    get:
      tags:
        - Vender
      summary: Ver detalle de solicitud
      description: |
        Según el estado muestra distinta info:
        - en_revision → Pantalla 49
        - aceptada → Pantalla 50 (valor base, comisiones, subasta asignada)
        - rechazada → Pantalla 52 (motivo rechazo, devolución con cargo)
        - en_subasta → Pantalla 54 (ubicación depósito, póliza seguro)
        - vendida → Pantalla 58
        - no_vendida → Pantalla 59 (la empresa compra por valor base)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalle de la solicitud según su estado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudVenta'
        '404':
          description: Solicitud no encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: SOLICITUD_NO_ENCONTRADA
                  message:
                    type: string
                    example: La solicitud de venta no existe o no pertenece a tu cuenta.
                  details:
                    type: object
                    nullable: true
                    example: null
...